chore: housekeeping across build, tasks, enums and tests
Directory.Packages.props declared Microsoft.Extensions.DependencyInjection,
.Hosting.Abstractions and .Logging, which no project references and which
were the only entries still pinned to 9.0.9.

Directory.Build.props set a non-existent <Author> property and applied
UserSecretsId to every project, including the packable library, which
baked a UserSecretsIdAttribute into the shipped assembly. The id now sits
on the two projects that read secrets.

The VS Code build task targeted the Release solution, watch ran the
ConsoleApp sample which now requires a media path argument, and there was
no test task.

The six library enums move into _Enums.cs, matching the Picker folder and
the repository convention.

Six dependency-injection tests lived in the rate-limiting and picker test
classes and move to ServiceCollectionExtensionsTests, three duplicated
stub handlers collapse into one shared type, and the two identical upload
error tests become a single theory.
