+++ Directory.Build.props
@@ -6,11 +6,6 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <!--nullable we disable on a per project basis-->
     <Nullable>enable</Nullable>
-    <UserSecretsId>bf9d717e-ecd3-40e4-850d-14010c167289</UserSecretsId>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <Author>Alex Vincent</Author>
   </PropertyGroup>
 
   <PropertyGroup>
