+++ src/CasCap.Api.GooglePhotos.Tests/Tests/StubHttpMessageHandler.cs
@@ -1,0 +1,17 @@
+namespace CasCap.Tests;
+
+/// <summary>Returns canned responses so tests can exercise HTTP code paths without a network.</summary>
+internal sealed class StubHttpMessageHandler : HttpMessageHandler
+{
+    private readonly Func<HttpRequestMessage, CancellationToken, Task<HttpResponseMessage>> _responseFactory;
+
+    internal StubHttpMessageHandler(Func<HttpRequestMessage, CancellationToken, Task<HttpResponseMessage>> responseFactory)
+        => _responseFactory = responseFactory;
+
+    internal StubHttpMessageHandler(Func<HttpRequestMessage, HttpResponseMessage> responseFactory)
+        => _responseFactory = (request, _) => Task.FromResult(responseFactory(request));
+
+    /// <inheritdoc />
+    protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        => _responseFactory(request, cancellationToken);
+}
