+++ src/CasCap.Api.GooglePhotos.Tests/Tests/Unit/GooglePhotosAuthorizationTests.cs
@@ -86,10 +86,4 @@ public sealed class GooglePhotosAuthorizationTests
         {
             BaseAddress = new Uri(RequestUris.BaseAddress)
         };
-
-    private sealed class StubHttpMessageHandler(Func<HttpRequestMessage, HttpResponseMessage> responseFactory) : HttpMessageHandler
-    {
-        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-            => Task.FromResult(responseFactory(request));
-    }
 }
