+++ src/CasCap.Api.GooglePhotos.Tests/Tests/Unit/GooglePhotosWriteRateLimitingHandlerTests.cs
@@ -29,103 +29,6 @@ public sealed class GooglePhotosWriteRateLimitingHandlerTests
     }
 
     [Fact]
-    public void RegistrationCopiesRateLimitOptions()
-    {
-        var configuredOptions = CreateEnabledOptions();
-        var services = new ServiceCollection();
-        services.AddGooglePhotos(new GooglePhotosOptions
-        {
-            User = "user@example.com",
-            ClientId = "client-id",
-            ClientSecret = "client-secret",
-            Scopes = [GooglePhotosScope.AppendOnly],
-            WriteRateLimit = configuredOptions
-        });
-        using var serviceProvider = services.BuildServiceProvider();
-        var resolvedOptions = serviceProvider.GetRequiredService<IOptions<GooglePhotosOptions>>().Value;
-
-        configuredOptions.PermitLimit = 2;
-
-        Assert.Equal(1, resolvedOptions.WriteRateLimit.PermitLimit);
-    }
-
-    [Fact]
-    public void RegistrationCopiesScopes()
-    {
-        GooglePhotosScope[] scopes = [GooglePhotosScope.AppendOnly];
-        var services = new ServiceCollection();
-        services.AddGooglePhotos(new GooglePhotosOptions
-        {
-            User = "user@example.com",
-            ClientId = "client-id",
-            ClientSecret = "client-secret",
-            Scopes = scopes
-        });
-        using var serviceProvider = services.BuildServiceProvider();
-        var resolvedOptions = serviceProvider.GetRequiredService<IOptions<GooglePhotosOptions>>().Value;
-
-        scopes[0] = GooglePhotosScope.PickerMediaItemsReadOnly;
-
-        Assert.Equal(GooglePhotosScope.AppendOnly, Assert.Single(resolvedOptions.Scopes));
-    }
-
-    [Fact]
-    public void RegistrationRejectsInvalidRateLimit()
-    {
-        var services = new ServiceCollection();
-        services.AddGooglePhotos(new GooglePhotosOptions
-        {
-            User = "user@example.com",
-            ClientId = "client-id",
-            ClientSecret = "client-secret",
-            WriteRateLimit = new GooglePhotosWriteRateLimitOptions
-            {
-                Enabled = true,
-                SegmentsPerWindow = 2,
-                WindowSeconds = 1
-            }
-        });
-        using var serviceProvider = services.BuildServiceProvider();
-
-        Assert.Throws<OptionsValidationException>(
-            () => serviceProvider.GetRequiredService<IOptions<GooglePhotosOptions>>().Value);
-    }
-
-    [Fact]
-    public void RegistrationRejectsNullRateLimit()
-    {
-        var services = new ServiceCollection();
-        services.AddGooglePhotos(new GooglePhotosOptions
-        {
-            User = "user@example.com",
-            ClientId = "client-id",
-            ClientSecret = "client-secret",
-            WriteRateLimit = null!
-        });
-        using var serviceProvider = services.BuildServiceProvider();
-
-        Assert.Throws<OptionsValidationException>(
-            () => serviceProvider.GetRequiredService<IOptions<GooglePhotosOptions>>().Value);
-    }
-
-    [Fact]
-    public void RegistrationRejectsNullScopes()
-    {
-        var services = new ServiceCollection();
-        services.AddGooglePhotos(new GooglePhotosOptions
-        {
-            User = "user@example.com",
-            ClientId = "client-id",
-            ClientSecret = "client-secret",
-            Scopes = null!
-        });
-        using var serviceProvider = services.BuildServiceProvider();
-
-        Assert.Throws<OptionsValidationException>(
-            () => serviceProvider.GetRequiredService<IOptions<GooglePhotosOptions>>().Value);
-    }
-
-    [Fact]
     public async Task QueuedWriteHonorsCancellation()
     {
         var requestCount = 0;
@@ -250,11 +153,5 @@ public sealed class GooglePhotosWriteRateLimitingHandlerTests
     {
         using var request = new HttpRequestMessage(method, $"https://photoslibrary.googleapis.com/v1/{requestUri}");
         return await client.SendAsync(request, cancellationToken);
-    }
-
-    private sealed class StubHttpMessageHandler(Func<HttpRequestMessage, HttpResponseMessage> responseFactory) : HttpMessageHandler
-    {
-        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-            => Task.FromResult(responseFactory(request));
     }
 }
