--- src/CasCap.Api.GooglePhotos/Models/GooglePhotosFeatureType.cs
@@ -1,10 +1,0 @@
-namespace CasCap.Models;
-
-/// <summary>Specifies a Google Photos feature used to filter media.</summary>
-[JsonConverter(typeof(JsonStringEnumConverter))]
-public enum GooglePhotosFeatureType
-{
-    /// <summary>Media marked as a favorite.</summary>
-    [JsonStringEnumMemberName("FAVORITES")]
-    Favorites
-}
